refactor: organize msw plugin files
