+++ dev/playground.ts
@@ -1,7 +1,11 @@
+/* eslint-disable */
 import type { DefinePlugin, IR } from '@hey-api/openapi-ts';
+import { http, HttpResponse } from 'msw';
+import { setupServer } from 'msw/node';
 
-// import { createOpencode } from '@opencode-ai/sdk';
 import { client } from './gen/typescript/client.gen';
+// import { createOpencode } from '@opencode-ai/sdk';
+import { createMswHandlerFactory, getAllMocks } from './gen/typescript/msw.gen';
 import { OpenCode } from './gen/typescript/sdk.gen';
 
 type MyPluginConfig = { readonly name: 'myplugin' };
@@ -16,6 +20,21 @@ export const handler: MyPlugin['Handler'] = ({ plugin }) => {
     console.log(event);
   });
 };
+
+const server = setupServer(
+  // ...getAllMocks({
+  //   overrides: {
+  //     // ...
+  //   },
+  // }),
+  createMswHandlerFactory({ baseUrl: '*' }).tuiPublishMock(),
+  // http.post('*/tui/publish', () => HttpResponse.json({
+  //   firstName: 'John',
+  //   id: 'abc-123',
+  //   lastName: 'Maverick',
+  // })),
+);
+server.listen();
 
 async function run() {
   // const { client, server } = await createOpencode();
